Added buttons to rows and created composable to delete document by id
