+++ vue-firebase-app/src/components/ManageClients.vue
@@ -29,6 +29,14 @@
           <td>{{ client.lastName }}</td>
           <td>{{ client.email }}</td>
           <td>{{ client.organization }}</td>
+          <td>
+            <button class="btn btn-primary btn-sm">
+              Modify
+            </button>
+            <button class="btn btn-warning btn-sm">
+              Delete
+            </button>
+          </td>
         </tr>
       </tbody>
     </table>
@@ -37,8 +45,9 @@
 </template>
 
 <script>
+  //import modifyDocument from '../composables/modifyDocument'
   import getClients from '../composables/getClients'
-  import modal from '../components/ClientEntry'
+  import modal from '../components/ClientModal'
   import { ref } from 'vue'
 
   export default {
@@ -48,7 +57,7 @@
       const showModal = ref(false)
       const currentPage = ref(1)
       const { clients, error, loadClients } = getClients()
-
+      //const { deleteDoc } = modifyDocument()
       // Loads the clients for the data table
       loadClients()
